Use connectbox jessie-backports repo for Raspbian (#91)
Use connectbox jessie-backports repo for Raspbian

The Debian Jessie-backports repo is armhf but compilied for ARMv7+ and
is unusable on the Pi Zero family so we introduce a repo that we
manage that provides just the packages we need from jessie-backports
compiled with the Raspbian toolchain as armhf for ARMv6+

The default values are for Debian/Armbian. If we ever get an Armbian
device with an ARMv6 CPU, then we'll need to split out Armbian and
our choice of defaults will no longer be so clear given the defaults
would only server a single platform.