Fix MacOS Yosemite captive portal behaviour (#94)
Yosemite and later have a different captive portal check to their
older MacOS brethren, so we need to add support for it.

This removes some user-agent logic from the nginx default server
that is no longer necessary. Previously the default server and the
connectbox server stanzas were combined, so we wanted to check the
useragent before responding with the magic captive portal responses
but now that it's separate, we don't have any risk of clashing with
a legit connectbox content request. This means we don't need the
user agent check. This meant that we could remove a negative test,
though we also added a test for Yosemite captive portal behaviour.