Workaround disappearing /var/log/nginx on armbian
There's a problem with the log2ram service when it attempts to flush
its contents to disk on shutdown. This manifests in nginx not
starting after a reboot because /var/log/nginx is missing. As a
workaround, we disabled log2ram at the start of the playbook run
so that the directory state on disk is what we want it to be. At
the end of the playbook run we re-enable it again.

https://forum.armbian.com/index.php?/topic/3729-problems-since-versions-525-and-526-with-varlog/

I'm unable to test this as I don't have any working Armbian devices
at the moment.