Re-licence under MIT licence (#109)
* Switch from Apache licence to MIT licence
* Mention third-party work

Move libraries and css under "external" directories so that we can
highlight where an alternative licence may apply. All files in an
external directory have a header with their licence.

Make specific mention of the licence for Font Awesome fonts as they
do not have a licence clearly visible in the files.