Set the locale in the admin api
Armbian and Raspbian set the locale appropriately, but the official
Ubuntu AMI that we'll use for CI does not, which means that
escapeshellcmd strips out unicode characters, which is problematic
given we explicitly test setting SSIDs containing unicode strings.

I don't think it's a problem to set this in the script, given we
control the connectbox environment and are shipping the device
with LC_CTYPE="en_US.UTF-8"  (it'd be LC_CTYPE="en_GB.UTF-8" if we
shipped a raspbian device), and both of those locales would allow
unicode characters through.