Add Ubuntu to CI pipeline (#135)
This adds Ubuntu 16.04 to our CI pipeline, alongside Debian Jessie.

As the official Ubuntu 16.04 AMI doesn’t ship with python3, we now
use Ansible 2.3 for deployment (which has support for python3 on
the remote side). There’s an ansible-lint version bump too because
the internals of Ansible changed between 2.1 and 2.3.

We now have a pre-constructed inventory file for ci and a simplified
build script. Aptitude is not present on Ubuntu AMIs, so it is also
added, even though it’s present on Armbian Ubuntu images.

Fixes #131