Resolved this (read more)
Pi has PHP 7.3 in apt already.
Debian 9 (AWS and ConnectBox) does not.  The Sury repo has PHP 7.3 but adding that repo fails on Pi so we are ingnoring the failure.

With all these, if PHP doesn't load this will fail so I'm ok ignoring the repo failures if they occur.