New README.txt for RPi builds
The handling of a mounted USB stick on an RPi build is different than that on a NEO build. Because of this, the default README.txt written for the NEO environment is incorrect. This change uses a new README.txt file for RPi builds which correctly informs the user of how media content is handled on the RPi.