Fix install problems with new RPi.GPIO
For some reason, the creation of the battery_tool_venv is installing an unwanted version of RPi.GPIO (version 0.7.0) which is only usable in a Raspbian environment. These changes delete that version of RPi.GPIO prior to installing the correct version. Also, we edit the /boot/armbianEnv.txt file to enable i2c0.